rag_app/pyproject.toml
[project]
name = "rag_app"
version = "0.1.0"
description = "a rag app powered by langchain and ollama"
authors = [
    {name = "Antonio Badilla-Olivas"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-ollama (>=0.2.2,<0.3.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "pypdf (>=5.2.0,<6.0.0)",
    "langchain-chroma (>=0.2.1,<0.3.0)",
    "langgraph (>=0.2.73,<0.3.0)",
    "langchain-text-splitters (>=0.3.6,<0.4.0)",
    "gradio (>=5.16.0,<6.0.0)"
]

[tool.poetry]
package-mode = false
